flash operation failed

+++ driver/hw_type.h
@@ -18,10 +18,10 @@ typedef enum
 } bStatus;
    
 /* common type defines */
-typedef unsigned int uint32;
+typedef unsigned long uint32;
 typedef unsigned short uint16;
 typedef unsigned char uint8;
-typedef signed int int32;
+typedef signed long int32;
 typedef signed short int16;
 typedef signed char int8;
 typedef enum bStatus
